requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
cloudinary==1.20.0
dj-database-url==0.5.0
Django==3.0.3
django-appconf==1.0.3
django-cloudinary-storage==0.2.3
django-comments-xtd==2.4.2
django-contrib-comments==1.9.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-js-asset==1.2.2
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
docutils==0.16
gunicorn==20.0.4
idna==2.9
jsmin==2.2.2
mysqlclient==1.4.6
pilkit==2.0
Pillow==7.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
